requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
passlib[bcrypt]
pydantic
alembic
python-dotenv
